all: refactor Update() and Layout() to support scroll containers
This is a big refactor that changes the signature and semantics of
Update() and Layout in a big way. While the source code gets bigger and
a bit more complex, the binary size is actually almost identical
(perhaps even slightly smaller) - probably because the compiler is
better able to const-propagate various values.

The refactor was needed:

  * To support scrollable views, which were hard to support in the
    previous design.
  * To fix some bugs that may have resulted in objects drawing outside
    the intended area.