ListBox: move parent scroll container into view
This is needed when a ListBox is made scrollable using a scroll
container and the user navigates using keys (instead of by scrolling).

It's not perfect: VBox currently doesn't propagate the 'scroll into
view' call and perhaps we should leave some space at the top/bottom if
that's available (like e.g. Canon cameras do in their menu) but it's way
better than not scrolling at all.